Created basics of ResourceManager
I have made the header of ResourceMangager, GUID and Resource.

The Meaning of Resource is it should be a parent class of all different kind of resources (mesh, texture, audo etc). I have choosen a dumb Resource architecture, where only the Manager can connect GUID's and paths. The Resource shall only be a dumnmy that holds data and an internal reference counter.